Move auth package to authority module
