--- shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/auth/AuthenticationContext.java
@@ -1,54 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.infra.metadata.auth;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * Authentication context.
- *
- * @author zhangliang
- */
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public final class AuthenticationContext {
-    
-    private static final AuthenticationContext INSTANCE = new AuthenticationContext();
-    
-    private volatile Authentication authentication;
-    
-    /**
-     * Get instance.
-     * 
-     * @return instance
-     */
-    public static AuthenticationContext getInstance() {
-        return INSTANCE;
-    }
-    
-    /**
-     * Initial authentication.
-     * 
-     * @param authentication authentication
-     */
-    public synchronized void init(final Authentication authentication) {
-        this.authentication = AuthenticationEngine.findSPIAuthentication().orElse(authentication);
-    }
-}
